MW-184 PR remarks
